Model's symmetry has been fixed, and added a pupil.
The model's symmetry should be correct now, considering that there is only
half of an object. I went through and made sure the 'midline' of the face
lines up correctly as well. Added a cylender cap which works as a "pupil"
and gave it a 'whiter' material than the eyeball, probably going to
inverse these materials.